add date and time picker ,implement and finish adding fragment using dagger2 and rx

+++ app/src/main/java/mobile/s/todogoals/database/ToDoDao.java
@@ -16,7 +16,7 @@ public interface ToDoDao {
     Flowable<List<ToDoModelDB>> getAllToDoList();
 
     @Query("DELETE FROM ToDoModelDB WHERE todoId= :id")
-    void deleteToDoById(Long id);
+    void deleteToDoById(int id);
 
     @Query("SELECT COUNT(todoId) FROM ToDoModelDB where todoDone=0")
     Flowable<Integer> getUnDoneToDoCount();
